requirements.txt
uvicorn
fastapi
pika
pyserial